Added some doc, and moved separated classes to files.
